[FSDP][state_dict][2/N] Move state_dict related enums/dataclasses/states to state_dict_utils.py and init_state_dict()
[ghstack-poisoned]
